requirements.txt
requests==2.32.3
Flask
scapy
python-dotenv
netifaces
